More renaming and add READMEs for CSV folders
