+++ Telemetry/telemetry_exe/Parsed_Data/README.md
@@ -1,0 +1,5 @@
+![HyTech Logo](https://hytechracing.gatech.edu/images/hytech_logo_small.png)
+# Raw Data Folder
+GET YOUR PARSED DATA CSVs HERE!!!
+This README is just a placeholder to ensure this folder is tracked after the global .gitignore.
+There is no need to delete CSVs here after each use - the global .gitignore untracks them.
